pyproject.toml
[project]
name = "watchmood-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.135.3",
    "google-genai>=1.73.1",
    "google-generativeai>=0.8.6",
    "pydantic>=2.13.0",
    "pydantic-ai>=1.82.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
